bodek-yang/pyproject.toml
[tool.poetry]
name = "bodek-yang"
version = "0.1.0"
description = ""
authors = ["Alessandro Candido <candido.ale@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10,<3.11"
h5py = "^3.7.0"
PyYAML = "^6.0"
yadism = {version = "^0.11.0", extras = ["mark"]}

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^8.4.0"
pdbpp = "^0.10.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
